Refactored int variables to float variables.

+++ CalculatorApp/Calculator/Program.cs
@@ -2,8 +2,8 @@
 
 using System.Threading.Channels;
 
-int num1 = 0;
-int num2 = 0;
+float num1 = 0;
+float num2 = 0;
 
 Console.WriteLine("Console Calculator in C#\r");
 Console.WriteLine("------------------------\n");
